fix CI error

+++ examples/shapes/shapes_dashed_line.c
@@ -79,7 +79,7 @@ int main(void)
             DrawText("LEFT/RIGHT: Change Space Length", 15, 55, 10, BLACK);
             DrawText("C: Cycle Color", 15, 75, 10, BLACK);
 
-            DrawText(TextFormat("Dash: %.0f | Space: %.0f", (int)dashLength, (int)blankLength), 15, 115, 10, DARKGRAY);
+            DrawText(TextFormat("Dash: %.0f | Space: %.0f", dashLength, blankLength), 15, 115, 10, DARKGRAY);
 
             DrawFPS(screenWidth - 80, 10);
